Debounce window resize callback for improved performance
Some browser events have an extremely high-frequency rate. The number of events such as cursor moving, scrolling, resizing the browser window, holding down a key, etc., can reach 100 times per second. Functions called at this frequency of events place a heavy load on the browser and it leads to a decrease in performance.
A debounce function will delay calling a function until after the specified delay has passed since the last time it was invoked. If the debounced function is invoked again before the delay has passed, it resets the delay.