Swap if conditions to profit from short-circuit evaluation on desktop
On mobile, the window rarely gets resized, therefore, we can swap !isNavbarHamburgerMenu() (always false on mobile) for isNavHamburgerMenuOpen() (50% chance of being true) since the event rarely triggers.
On desktop, !isNavbarHamburgerMenu() has 60% chance of being true and if true, will then evaluate isNavHamburgerMenuOpen() which always has 50% chance of being true.
Therefore, it's better to swap the conditions (if my math and logic are correct 🙃) (not that it matters in the slightest since performance improvements like this are negligible)